requirements.txt
altair==5.5.0
attrs==24.3.0
bcrypt==4.1.2
beautifulsoup4==4.12.3
blinker==1.9.0
cachetools==5.5.1
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
contourpy==1.3.1
cycler==0.12.1
dnspython==2.7.0
ecdsa==0.19.0
entrypoints==0.4
Faker==34.0.2
favicon==0.7.0
filelock==3.17.0
fonttools==4.55.4
fsspec==2024.12.0
gitdb==4.0.12
GitPython==3.1.44
htbuilder==0.9.0
huggingface-hub==0.27.1
idna==3.10
Jinja2==3.1.5
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kiwisolver==1.4.8
lxml==5.3.0
Markdown==3.7
markdown-it-py==3.0.0
markdownlit==0.0.7
MarkupSafe==3.0.2
matplotlib==3.10.0
mdurl==0.1.2
mpmath==1.3.0
narwhals==1.23.0
networkx==3.2.1
numpy==1.26.4
packaging==23.2
pandas==2.2.0
pillow==10.4.0
plotly==5.18.0
protobuf==4.25.5
pyarrow==19.0.0
pyasn1==0.6.1
pydeck==0.9.1
Pygments==2.19.1
pymdown-extensions==10.14.1
pymongo==4.6.1
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.3.0
pytz==2024.2
PyYAML==6.0.2
referencing==0.36.1
regex==2024.11.6
requests==2.31.0
rich==13.9.4
rpds-py==0.22.3
rsa==4.9
safetensors==0.5.2
six==1.17.0
smmap==5.0.2
soupsieve==2.6
st-annotated-text==4.0.1
streamlit==1.32.0
streamlit-camera-input-live==0.2.0
streamlit-card==1.0.2
streamlit-embedcode==0.1.2
streamlit-extras==0.3.6
streamlit-faker==0.0.3
streamlit-image-coordinates==0.1.9
streamlit-keyup==0.3.0
streamlit-lottie==0.0.5
streamlit-toggle-switch==1.0.2
streamlit-vertical-slider==2.5.5
sympy==1.13.1
tenacity==8.5.0
tokenizers==0.15.2
toml==0.10.2
torch==2.2.0+cpu
tornado==6.4.2
tqdm==4.67.1
transformers==4.36.2
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
watchdog==6.0.0
